# Docs/API_Endpoints/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    AccountResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        loginNr:
          type: string
        type:
          $ref: "#/components/schemas/AccountType"
    AccountType:
      type: string
      enum:
      - ADMIN
      - FOOD_COURT_WORKER
      - GUEST
...
    LoginRequest:
      type: object
      properties:
        loginNr:
          type: string
        password:
          type: string
...
    RegisterRequest:
      type: object
      properties:
        loginNr:
          type: string
        password:
          type: string
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
  /account/list_all:
    get:
      summary: List all Accounts
      responses:
        "200":
          description: List of Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Account Api
      security:
      - SecurityScheme:
        - ADMIN
  /account/login:
    post:
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Login succeeded; cookie set
          content:
            text/plain: {}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
      tags:
      - Account Api
  /account/logout:
    post:
      summary: Logout
      responses:
        "204":
          description: Logout succeeded; access token cleared
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Account Api
      security:
      - SecurityScheme:
        - GUEST
        - FOOD_COURT_WORKER
        - ADMIN
  /account/register:
    post:
      summary: Register a new Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  loginNr:
                    type: string
                  type:
                    $ref: "#/components/schemas/AccountType"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
      tags:
      - Account Api
...
info:
  title: App API
  version: 0.1.0-SNAPSHOT